+++ src/api/interfaces/CRUDApi.ts
@@ -1,5 +1,7 @@
 export default interface CRUDApi {
 
     fetchAll(): Promise<String>;
+
+    post(data: Object): Promise<Object|String>;
     
 }